# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Result:
    properties:
      course:
        type: string
      createdAt:
        type: string
      credit:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      grade:
        type: string
      id:
        type: integer
      semester:
        type: string
      studentID:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Student:
    properties:
      address:
        type: string
      admission_date:
        type: string
      batch:
        type: string
      blood_group:
        type: string
      cgpa:
        type: number
      createdAt:
        type: string
      credit_completed:
        type: integer
      current_semester:
        type: string
      date_of_birth:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      department:
        type: string
      email:
        type: string
      gender:
        type: string
      guardian_name:
        type: string
      guardian_phone:
        type: string
      guardian_relation:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      payment_status:
        type: string
      phone:
        type: string
      program:
        type: string
      results:
        items:
          $ref: '#/definitions/models.Result'
        type: array
      student_id:
        type: integer
      total_credits:
        type: integer
      total_due:
        type: number
      total_paid:
        type: number
      total_payable:
        type: number
      updatedAt:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for the Student Portal project.
  title: Student Portal API
  version: "1.0"
paths:
  /results:
    get:
      description: Retrieve all academic results from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Result'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all results
      tags:
      - Results
    post:
      consumes:
      - application/json
      description: Create a new academic result for a specific student
      parameters:
      - description: Result Info
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/models.Result'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Result'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new result for a student
      tags:
      - Results
  /students:
    get:
      description: Retrieve all students from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Create a new student with the provided information
      parameters:
      - description: Student Info
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new student
      tags:
      - Students
  /students/{id}:
    delete:
      description: Delete a student from the database
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a student by ID
      tags:
      - Students
    get:
      description: Retrieve a specific student by their ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a student by ID
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: Update an existing student's information
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated student data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a student by ID
      tags:
      - Students
  /students/{id}/results:
    get:
      description: Retrieve all academic results for a specific student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Result'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get results for a specific student
      tags:
      - Results
schemes:
- http
swagger: "2.0"